# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Learning Time Tracker API
  version: '1'
  description: >
    API-first backend for Learning Time Tracker. FE never talks to the
    database or file storage directly — every read/write goes through this
    API. Auth: FE signs in via Firebase client SDK (email+password or
    Google); every protected endpoint requires a Firebase ID token in
    `Authorization: Bearer <token>`, verified server-side via Firebase Admin
    SDK.
servers:
  - url: /api/v1
paths:
...
  /users/me/competencies:
    get:
      summary: List the caller's own competencies
      description: >
        AUTH-005/AUTH-006/AUTH-007. Dictionary picks and free-text (custom)
        entries alike, each with its own `willingToTeach` flag. Always scoped
        to the caller (`req.firebaseUser.uid`) — there is no way to list
        another user's competencies through this API.
      operationId: listMyCompetencies
      tags: [profile]
      security:
        - firebaseIdToken: []
      responses:
        '200':
          description: The caller's competencies
          content:
            application/json:
              schema:
                type: object
                properties:
                  competencies:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserCompetency'
                required: [competencies]
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      summary: Add a competency to the caller's profile
      description: >
        Exactly one of `competencyId` (a dictionary pick, AUTH-005) or
        `customLabel` (free text, AUTH-006) must be given — both or neither
        is 400 `errors.competency.invalidPayload`.


        **AUTH-005 AC3/AC5**: re-adding a `competencyId` already on the
        caller's profile is 409 `errors.competency.alreadyAdded` (no
        upsert/idempotency here, unlike board_members/task_shares) — an
        unknown or inactive `competencyId` is 400
        (`errors.competency.notFound` / `errors.competency.inactive`
        respectively). A blank/whitespace-only `customLabel` is 400
        `errors.competency.customLabelRequired`; longer than 100 characters is
        `errors.competency.customLabelTooLong`. Duplicate `customLabel`
        values ARE allowed (AUTH-006 AC7) — free text is never deduplicated.
      operationId: addMyCompetency
      tags: [profile]
      security:
        - firebaseIdToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCompetencyCreateRequest'
      responses:
        '201':
          description: The created competency
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserCompetency'
        '400':
          description: >
            Invalid payload shape, unknown/inactive competencyId, or invalid
            customLabel — see the messageKey table above.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: This `competencyId` is already on the caller's profile (`errors.competency.alreadyAdded`)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          $ref: '#/components/responses/InternalError'
  /users/me/competencies/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    patch:
      summary: Toggle a competency's "willing to teach" flag
      description: >
        AUTH-007. Works for both dictionary and custom competencies (AC4) —
        the flag doesn't depend on origin.


        **Anti-enumeration (AC3)**: an id belonging to another user's
        `user_competencies` row resolves to the SAME 404
        `errors.competency.notFound` as a genuinely missing id — never 403.
        Same convention already applied to `time_entries` (US-011 AC7) and
        to sign-in (AUTH-002 AC2).
      operationId: updateMyCompetencyWillingToTeach
      tags: [profile]
      security:
        - firebaseIdToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCompetencyWillingToTeachRequest'
      responses:
        '200':
          description: The updated competency
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserCompetency'
        '400':
          description: '`willingToTeach` missing or not a boolean (`errors.competency.invalidPayload`)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: No such competency on the caller's own profile (see anti-enumeration note above)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      summary: Remove a competency from the caller's profile
      description: >
        AUTH-005 AC4 / AUTH-006 AC6 — a single endpoint for both dictionary
        and custom competencies. Removing a competency that has
        `willingToTeach = true` simply removes the row; the flag has no
        existence independent of it. Same anti-enumeration contract as PATCH
        above.
      operationId: removeMyCompetency
      tags: [profile]
      security:
        - firebaseIdToken: []
      responses:
        '204':
          description: Removed
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: No such competency on the caller's own profile (see anti-enumeration note above)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /competencies:
    get:
      summary: The active competencies dictionary
      description: >
        AUTH-005 AC1 — what the profile screen's picker offers. A competency
        retired via `is_active = false` (admin-managed, outside this API)
        simply stops appearing here; any `user_competencies` rows already
        referencing it are unaffected (AUTH-005 AC6).
      operationId: listCompetenciesCatalog
      tags: [profile]
      security:
        - firebaseIdToken: []
      responses:
        '200':
          description: The active dictionary, ordered by slug
          content:
            application/json:
              schema:
                type: object
                properties:
                  competencies:
                    type: array
                    items:
                      $ref: '#/components/schemas/Competency'
                required: [competencies]
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /competencies/{id}/chat/messages:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: A competency's group chat history
      description: >
        US-028 AC2. There is no separate "rooms" resource — a room IS a
        `competencies` row, identified directly by `id`. Any authenticated
        user can read regardless of whether this competency is in their own
        profile or `willing_to_teach` (decision #4 in this project's
        US-025…029 origin notes — no membership concept at all here, unlike
        every other resource in this API). Chronological, oldest first, no
        pagination in MVP. A retired (`is_active = false`) or nonexistent
        competency 404s — deactivation closes the product path but never
        deletes the room's history (AC5).
      operationId: listCompetencyChatMessages
      tags: [chat]
      security:
        - firebaseIdToken: []
      responses:
        '200':
          description: The room's messages, oldest first
          content:
            application/json:
              schema:
                type: object
                properties:
                  messages:
                    type: array
                    items:
                      $ref: '#/components/schemas/CompetencyChatMessage'
                required: [messages]
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: No such competency, or it has been retired (`errors.competencyChat.notFound`)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      summary: Post a message to a competency's group chat
      description: >
        US-028 AC3/AC4. Same "any authenticated user, no membership check"
        access as GET above. On success, a `competencyChat.message.created`
        WS event (see the "WebSocket channel" section near the end of this
        document) is pushed to every live connection currently subscribed to
        this room — REST remains the source of truth regardless of whether
        any such connection exists.
      operationId: createCompetencyChatMessage
      tags: [chat]
      security:
        - firebaseIdToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                body:
                  type: string
                  minLength: 1
                  maxLength: 2000
              required: [body]
      responses:
        '201':
          description: The newly created message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompetencyChatMessage'
        '400':
          description: >
            Empty/whitespace-only body (`errors.competencyChat.messageBodyRequired`)
            or body over 2000 characters (`errors.competencyChat.messageBodyTooLong`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: No such competency, or it has been retired (`errors.competencyChat.notFound`)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /dm-threads:
    post:
      summary: Get-or-create a DM thread with another user about one competency
      description: >
        US-027 AC1-3. A top-level resource, not nested under `/users` — a
        thread belongs to a PAIR of users plus a competency, with no single
        natural parent to nest it under (the group chat under
        `/competencies/{id}/chat` does have one, hence the different
        nesting there).


        Idempotent get-or-create: the SAME (caller, `targetUserId`,
        `competencyId`) always resolves to the SAME thread — uniqueness is
        enforced at the DB level on the normalized pair, so this is safe
        under concurrent calls (AC1). Returns `200` if the thread already
        existed, `201` if this call just created it — same JSON shape
        either way.


        `competencyId` must be one `targetUserId` has marked
        `willing_to_teach = true` for (determined by US-026 AC3-5's profile
        flow) — re-validated server-side on every call, including a direct
        API call that tries to bypass that UI selection (AC3).
      operationId: getOrCreateDmThread
      tags: [chat]
      security:
        - firebaseIdToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                targetUserId:
                  type: string
                  description: Firebase UID of the person to message
                competencyId:
                  type: string
                  format: uuid
              required: [targetUserId, competencyId]
      responses:
        '200':
          description: The existing thread for this pair+competency
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DmThread'
        '201':
          description: A newly created thread
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DmThread'
        '400':
          description: >
            `targetUserId` equal to the caller (`errors.dmThread.cannotMessageSelf`),
            or `competencyId` not marked `willing_to_teach = true` by
            `targetUserId` — including a nonexistent/malformed id, which is
            indistinguishable from "not offered" (`errors.dmThread.competencyNotOffered`),
            or a missing/malformed `targetUserId` (`errors.dmThread.invalidPayload`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    get:
      summary: List my DM threads
      description: >
        US-027 AC11/AC12. Every thread the caller participates in, sorted by
        last activity (most recent message, or the thread's own
        `createdAt` if it has no messages yet, AC12) descending. No
        pagination in MVP.
      operationId: listMyDmThreads
      tags: [chat]
      security:
        - firebaseIdToken: []
      responses:
        '200':
          description: The caller's threads, most recently active first
          content:
            application/json:
              schema:
                type: object
                properties:
                  threads:
                    type: array
                    items:
                      $ref: '#/components/schemas/DmThread'
                required: [threads]
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /dm-threads/{id}/messages:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: A DM thread's message history
      description: >
        US-027 AC4/AC5. Participants only — DM privacy is absolute, there is
        no owner/admin exception anywhere in this project for someone else's
        private messages. A nonexistent thread id and "thread exists but I'm
        not a participant" are BOTH `403 errors.dmThread.forbidden` — never a
        distinct 404 — so a thread id alone never reveals whether it names a
        real conversation between two other people (same anti-enumeration
        principle the WS `subscribe` frame's authorization check applies,
        see the "WebSocket channel" section near the end of this document).
        Chronological, oldest first, no pagination in MVP.
      operationId: listDmThreadMessages
      tags: [chat]
      security:
        - firebaseIdToken: []
      responses:
        '200':
          description: The thread's messages, oldest first
          content:
            application/json:
              schema:
                type: object
                properties:
                  messages:
                    type: array
                    items:
                      $ref: '#/components/schemas/DmMessage'
                required: [messages]
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Not a participant of this thread, or the thread doesn't exist (`errors.dmThread.forbidden`)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      summary: Send a message in a DM thread
      description: >
        US-027 AC5-8. Same participants-only gate as GET above. On success,
        a `dm.message.created` WS event (see the "WebSocket channel" section
        near the end of this document) is pushed to the other participant's
        live connection(s), if any, AND to any connection subscribed to this
        specific thread's channel — REST remains the source of truth
        regardless of whether any such connection exists (AC8: a missing/
        dropped WS connection never loses the message).
      operationId: createDmThreadMessage
      tags: [chat]
      security:
        - firebaseIdToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                body:
                  type: string
                  minLength: 1
                  maxLength: 2000
              required: [body]
      responses:
        '201':
          description: The newly created message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DmMessage'
        '400':
          description: >
            Empty/whitespace-only body (`errors.dmThread.messageBodyRequired`)
            or body over 2000 characters (`errors.dmThread.messageBodyTooLong`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Not a participant of this thread, or the thread doesn't exist (`errors.dmThread.forbidden`)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  responses:
    Unauthorized:
      description: Missing, malformed, or invalid/expired Firebase ID token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  securitySchemes:
    firebaseIdToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase ID token, verified via Firebase Admin SDK.
  schemas:
...
    Competency:
      type: object
      description: One entry in the active competencies dictionary (AUTH-005).
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
          description: Locale lookup key — rendered via `competency.<slug>` in the FE's own dictionary, never returned pre-translated.
      required: [id, slug]
    UserCompetency:
      type: object
      description: >
        One of the caller's own competencies — either a dictionary pick
        (`isCustom = false`, `competencyId`/`competencySlug` set) or a
        free-text entry (`isCustom = true`, `customLabel` set). Exactly one
        of `competencySlug`/`customLabel` is non-null on any given row.
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          description: Firebase UID — always the caller for any row this API ever returns.
        competencyId:
          type: string
          format: uuid
          nullable: true
        competencySlug:
          type: string
          nullable: true
          description: Set only when isCustom is false — render via `competency.<slug>`, same as Competency.slug above.
        isCustom:
          type: boolean
        customLabel:
          type: string
          nullable: true
          description: Set only when isCustom is true. User-generated text — rendered as-is, never run through the locale dictionary.
        willingToTeach:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, userId, competencyId, competencySlug, isCustom, customLabel, willingToTeach, createdAt, updatedAt]
    UserCompetencyCreateRequest:
      type: object
      description: Exactly one of `competencyId`/`customLabel` — see POST /users/me/competencies for the full validation rules.
      properties:
        competencyId:
          type: string
          format: uuid
        customLabel:
          type: string
          maxLength: 100
    UserCompetencyWillingToTeachRequest:
      type: object
      properties:
        willingToTeach:
          type: boolean
      required: [willingToTeach]
...
    DmThreadOtherUser:
      type: object
      properties:
        id:
          type: string
          description: Firebase UID of the OTHER participant (never the caller).
        name:
          type: string
          nullable: true
    DmThreadLastMessage:
      type: object
      nullable: true
      description: Null for a freshly-created thread with no messages yet (US-027 AC12).
      properties:
        body:
          type: string
        createdAt:
          type: string
          format: date-time
    DmThread:
      type: object
      description: A `dm_threads` row, hydrated for display (US-027 AC1/AC11).
      properties:
        id:
          type: string
          format: uuid
        competencyId:
          type: string
          format: uuid
        competencySlug:
          type: string
          nullable: true
        otherUser:
          $ref: '#/components/schemas/DmThreadOtherUser'
        lastMessage:
          $ref: '#/components/schemas/DmThreadLastMessage'
        createdAt:
          type: string
          format: date-time
      required: [id, competencyId, competencySlug, otherUser, lastMessage, createdAt]
    DmMessage:
      type: object
      description: >
        A `dm_messages` row (US-027). View + add only — no `updatedAt`
        (same MVP scope decision as `TaskComment`/`CompetencyChatMessage`).
      properties:
        id:
          type: string
          format: uuid
        threadId:
          type: string
          format: uuid
        senderId:
          type: string
          description: Firebase UID of the sender.
        senderName:
          type: string
          description: AUTH-004 AC5/AC6 publicName-falls-back-to-displayName pattern.
        body:
          type: string
          maxLength: 2000
        createdAt:
          type: string
          format: date-time
      required: [id, threadId, senderId, senderName, body, createdAt]
    CompetencyChatMessage:
      type: object
      description: >
        A `competency_chat_messages` row (US-028) — the shared room
        identified by `competencyId` (no separate "rooms" resource). View +
        add only, same MVP scope as `DmMessage`/`TaskComment`.
      properties:
        id:
          type: string
          format: uuid
        competencyId:
          type: string
          format: uuid
        senderId:
          type: string
          description: Firebase UID of the sender.
        senderName:
          type: string
          description: AUTH-004 AC5/AC6 publicName-falls-back-to-displayName pattern.
        body:
          type: string
          maxLength: 2000
        createdAt:
          type: string
          format: date-time
      required: [id, competencyId, senderId, senderName, body, createdAt]
    ApiError:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: AUTH_TOKEN_INVALID
            messageKey:
              type: string
              description: Localization key the FE looks up in its own locale dictionary.
              example: errors.auth.tokenInvalid
            message:
              type: string
              description: English fallback for logs/devtools — not shown to end users.
          required: [code, messageKey]
      required: [error]
security:
  - firebaseIdToken: []

# ---------------------------------------------------------------------------
# WebSocket channel (US-029) — documented here as plain text, NOT as an
# OpenAPI `paths` entry, since OpenAPI 3.0 has no syntax for a WS
# upgrade/frame protocol. This is the one deliberate, explicitly-documented
# departure from "every endpoint is a REST path in this file" (CLAUDE.md
# "Архітектура": the FE ↔ BE contract was pure REST before US-029).
#
# Placement decision (US-029 AC6/AC7): the WS server is attached to the
# SAME http.Server/port (4000) the Express app already listens on — see
# backend/src/ws/server.js's header comment for the full rationale and
# trade-offs. `docker-compose.yml` is therefore UNCHANGED by this pass;
# there is no new service to add.
#
# Endpoint:      ws(s)://<same host:port as the REST API>/ws?token=<Firebase ID token>
#
# Authentication (US-029 AC2/AC3): the SAME Firebase ID token used for
# `Authorization: Bearer` on REST calls, passed as the `token` query
# parameter (the browser's native WebSocket API can't set custom headers on
# the upgrade request). Verified via the same Firebase Admin SDK
# `verifyIdToken()` the REST `requireAuth` middleware uses — not a separate
# auth mechanism. A missing or invalid/expired token gets a structured
# `error` frame (`ws.error.unauthorized`) followed immediately by the
# server closing the connection with close code 4401 — no subscription or
# message is ever accepted first.
#
# Delivery guarantee (US-027 AC8/AC9, US-029 AC5/AC9): REST is unconditionally
# the source of truth. Every message is written to Postgres by a REST POST
# BEFORE any WS event is broadcast; a missing/dropped/never-subscribed WS
# connection never loses a message — the full history is always available
# via `GET /dm-threads/{id}/messages` / `GET /competencies/{id}/chat/messages`.
# WS only accelerates "live" delivery to connections that happen to be
# listening at that instant.
#
# --- Client -> Server frames ---
#
#   {"type": "subscribe", "channel": "dmThread", "threadId": "<uuid>"}
#     Authorized via the SAME service-layer check
#     (`requireDmThreadAccess`) `GET /dm-threads/{id}/messages` uses — only
...